backend/requirements.txt
fastapi
python-multipart
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
jinja2
requests
psycopg2-binary
python-dotenv
PyYAML 